backend/.history/requirements_20250714083743.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
pymysql
pandas
numpy
joblib
xgboost
openpyxl
